fix up url
